requirements.txt
fastapi>=0.111
uvicorn[standard]
pydantic>=2.7
python-multipart
sqlalchemy>=2.0
asyncpg
alembic
boto3
redis>=5.0
celery[redis]>=5.4
jinja2
python-docx
docxtpl
pydantic-settings
requests
mammoth
html2docx
slowapi
pytesseract
Pillow

PyJWT

bcrypt

structlog

prometheus-fastapi-instrumentator

psycopg2-binary

python-magic
